+++ src/main/java/com/kard/api/resources/users/rewards/types/PlacementBatchAttributes.java
@@ -12,6 +12,7 @@ import com.fasterxml.jackson.annotation.JsonSetter;
 import com.fasterxml.jackson.annotation.Nulls;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.kard.api.core.ObjectMappers;
+import com.kard.api.resources.commons.types.OfferComponents;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;